+++ trunk/source/OSXUI/HashBridge.h
@@ -18,6 +18,7 @@
 - (void)clear;
 
 - (void)setStop:(bool)val;
+- (void)setPause:(bool)val;
 - (void)setUppercase:(bool)val;
 - (int)getProgMax;
 - (uint64_t)getTotalSize;